requirements.txt
# requirements.txt - Залежності проекту

# Основні залежності
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0

# Розробка та тестування (опціонально)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Форматування коду (опціонально)
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Моніторинг помилок (опціонально, для продакшену)
# sentry-sdk==1.39.2